refactor-#164: graphql mutations and queries for payment
