+++ src/microservices/payment/graphql/mutations/index.ts
@@ -1,0 +1,1 @@
+export * from './payment-mutation.resolver';
