+++ src/microservices/payment/graphql/mutations/payment-mutation.resolver.ts
@@ -2,9 +2,9 @@ import { Args, Mutation, Resolver } from '@nestjs/graphql';
 import { UseGuards } from '@nestjs/common';
 
 import { GqlJwtAuthGuard } from '@auth/graphql/guards';
-import { GqlPayment } from './models';
-import { CreatePaymentInput } from './input';
-import { PaymentService } from '../services';
+import { GqlPayment } from '../models';
+import { CreatePaymentInput } from '../input';
+import { PaymentService } from '../../services';
 
 @Resolver((of) => GqlPayment)
 export class PaymentMutationResolver {
