+++ src/microservices/payment/graphql/queries/index.ts
@@ -1,0 +1,1 @@
+export * from './payment-query.resolver';
