+++ src/microservices/payment/graphql/queries/payment-query.resolver.ts
@@ -2,8 +2,8 @@ import { UseGuards } from '@nestjs/common';
 import { Args, Resolver, Int, Query } from '@nestjs/graphql';
 
 import { GqlJwtAuthGuard } from '@auth/graphql/guards';
-import { PaymentService } from '../services';
-import { GqlPayment } from './models/payment-gql.model';
+import { PaymentService } from '../../services';
+import { GqlPayment } from '../models';
 
 @Resolver((of) => GqlPayment)
 export class PaymentQueryResolver {
