+++ src/microservices/payment/payment.module.ts
@@ -6,8 +6,8 @@ import { BookingModule } from '@booking/booking.module';
 import { BookingService } from '@booking/services';
 import { PaymentService } from './services';
 import { PaymentController } from './rest/controllers';
-import { PaymentMutationResolver } from './graphql/payment-mutation.resolver';
-import { PaymentQueryResolver } from './graphql/payment-query.resolver';
+import { PaymentMutationResolver } from './graphql/mutations';
+import { PaymentQueryResolver } from './graphql/queries';
 
 @Module({
   imports: [
